[Travis] Fix phan errors on the release branch

+++ modules/issue_tracker/ajax/EditIssue.php
@@ -149,20 +149,22 @@ function editIssue()
     }
 
     // Add editor to the watching table unless they don't want to be added.
-    if ($_POST['watching'] == 'Yes') {
-        $nowWatching = array(
-                        'userID'  => $user->getData('UserID'),
-                        'issueID' => $issueID,
-                       );
-        $db->replace('issues_watching', $nowWatching);
-    } else if ($_POST['watching'] == 'No') {
-        $db->delete(
-            'issues_watching',
-            array(
-             'issueID' => $issueID,
-             'userID'  => $user->getData('UserID'),
-            )
-        );
+    if (isset($_POST['watching'])) {
+        if ($_POST['watching']== 'Yes') {
+            $nowWatching = array(
+                'userID'  => $user->getData('UserID'),
+                'issueID' => $issueID,
+            );
+            $db->replace('issues_watching', $nowWatching);
+        } else if ($_POST['watching'] == 'No') {
+            $db->delete(
+                'issues_watching',
+                array(
+                    'issueID' => $issueID,
+                    'userID'  => $user->getData('UserID'),
+                )
+            );
+        }
     }
     return ['issueID' => $issueID];
 }
